day 3, rename classes
