+++ src/main/java/org/example/fabricOrder/BuilderOrder.java
@@ -5,6 +5,8 @@ import org.example.order.Order;
 
 public class BuilderOrder {
 
+    Order order;
+
