+++ src/main/java/org/example/main/Main.java
@@ -1,12 +1,12 @@
 package org.example.main;
 
 import org.example.enums.EnumCourier;
-import org.example.fabric.Builder;
+import org.example.fabricPerson.BuilderPerson;
 import org.example.person.Person;
 
 public class Main {
     public static void main(String[] args) throws Exception {
-        Person car = new Builder(EnumCourier.CAR).build();
+        Person car = new BuilderPerson(EnumCourier.CAR).build();
         System.out.println(car.toString());
         System.out.println(car.toString());
     }
